Optimize compilation time for binius_field tests (#287)
### TL;DR

Move byte-sliced tests from module code to integration tests to improve compilation time. On my laptop this reduces compilation time of `binius_field(tests)` twice

### What changed?

This PR moves the byte-sliced field tests from the module code in `crates/field/src/arch/portable/byte_sliced/mod.rs` to a dedicated integration test file at `crates/field/tests/byte_sliced.rs`. The test logic remains the same, but is now properly organized as integration tests.

Key changes:
- Removed the `tests` module from the byte-sliced implementation file
- Created a new integration test file with the same test coverage
- Updated imports to use the public API through `binius_field` crate
- Replaced proptest-based test generation with standard test functions using random inputs

### How to test?

Run the test suite to verify all byte-sliced tests still pass:
```
cargo test -p binius-field
```

### Why make this change?

Moving these tests to integration tests reduces the compilation time of the field crate. Integration tests are compiled separately from the main crate, which means changes to the tests don't trigger recompilation of the entire crate. This improves the development experience by making the build process faster, especially during iterative development.
